Update: load initial post data
