+++ src/entities/labels/api/index.ts
@@ -1,0 +1,2 @@
+export { fetchLabels } from './fetch-labels';
+export { useFetchLabels } from './queries';
