+++ src/entities/labels/index.ts
@@ -1,0 +1,2 @@
+export { fetchLabels } from './api/fetch-labels';
+export { useFetchLabels } from './api/queries';
