+++ src/features/admin/edit-post-form/ui/edit-post-form-container.tsx
@@ -3,6 +3,7 @@ import { MultiValue } from 'chakra-react-select';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
+import { useFetchLabels } from '@entities/labels';
 import { usePostForm } from '@entities/posts/api/use-post-form';
 import { PostSchema } from '@entities/posts/model/schema';
 import { PostStatus } from '@entities/posts/model/types';
@@ -46,6 +47,7 @@ export const EditPostFormContainer = ({ slug }: EditPostFormContainerProps) => {
 
   const navigate = useNavigate();
   const { mutateAsync: submitPost } = useSubmitPost();
+  const { data: labels } = useFetchLabels();
   const onSubmitPost = async (data: PostSchema) => {
     await submitPost(data);
     navigate(paths.admin.dashboard);
@@ -66,7 +68,12 @@ export const EditPostFormContainer = ({ slug }: EditPostFormContainerProps) => {
     setValue('content', text);
   };
 
-  const labelOptions = data?.labels?.map((label) => ({
+  const labelOptions = labels?.map((label) => ({
+    value: label.id,
+    label: label.content,
+  }));
+
+  const selectedLabelOptions = data?.labels?.map((label) => ({
     value: label.id,
     label: label.content,
   }));
@@ -85,6 +92,8 @@ export const EditPostFormContainer = ({ slug }: EditPostFormContainerProps) => {
       handleContentChange={handleContentChange}
       watch={watch}
       setValue={setValue}
+      initialContentValues={data?.content}
+      selectedLabelOptions={selectedLabelOptions}
     />
   );
 };
